refactor: update TagSelector style and code structure
